requirements.txt
psycopg2-binary==2.9.3
psycopg==3.0.12
Flask == 2.2.2
